添加 发货管理 明细

+++ src/main/java/com/xiliulou/afterserver/service/impl/PointNewServiceImpl.java
@@ -338,7 +338,7 @@ public class PointNewServiceImpl extends ServiceImpl<PointNewMapper, PointNew> i
         return R.ok();
     }*/
 
-    @Override
+    /*@Override
     @Transactional(readOnly = false, propagation = Propagation.REQUIRED)
     public R deleteProduct(Long pointId, Long producutId) {
         if(pointId == null){
@@ -350,12 +350,12 @@ public class PointNewServiceImpl extends ServiceImpl<PointNewMapper, PointNew> i
 
         ProductNew productNew = productNewService.queryByIdFromDB(producutId);
         if(ObjectUtils.isNotNull(productNew)){
-            /*if(productNew.getCache() != null){
+            if(productNew.getCache() != null){
                 productNewMapper.delete(new UpdateWrapper<ProductNew>().eq("id", producutId));
             }else{
                 LambdaUpdateWrapper<File> wrapper = new LambdaUpdateWrapper<File>().eq(File::getType, File.TYPE_PRODUCT).eq(File::getBindId, producutId);
                 fileMapper.delete(wrapper);
-            }*/
+            }
 
             UpdateWrapper<PointProductBind> wrapper = new UpdateWrapper<>();
             wrapper.eq("product_id", producutId);
